+++ src/main/java/io/lettuce/core/api/reactive/RedisStreamReactiveCommands.java
@@ -143,9 +143,9 @@ public interface RedisStreamReactiveCommands<K, V> {
      *
      * @param key the stream key.
      * @param consumer consumer identified by group name and consumer key.
-     * @return simple-reply {@code true} if successful.
+     * @return Long integer-reply number of pending messages.
      */
-    Mono<Boolean> xgroupDelconsumer(K key, Consumer<K> consumer);
+    Mono<Long> xgroupDelconsumer(K key, Consumer<K> consumer);
 
     /**
      * Destroy a consumer group.
@@ -324,5 +324,4 @@ public interface RedisStreamReactiveCommands<K, V> {
      * @return simple-reply number of removed entries.
      */
     Mono<Long> xtrim(K key, boolean approximateTrimming, long count);
-
 }