+++ src/main/templates/io/lettuce/core/api/RedisStreamCommands.java
@@ -146,9 +146,9 @@ public interface RedisStreamCommands<K, V> {
      *
      * @param key the stream key.
      * @param consumer consumer identified by group name and consumer key.
-     * @return simple-reply {@code true} if successful.
+     * @return Long integer-reply number of pending messages.
      */
-    Boolean xgroupDelconsumer(K key, Consumer<K> consumer);
+    Long xgroupDelconsumer(K key, Consumer<K> consumer);
 
     /**
      * Destroy a consumer group.